refactor: move Co2Forecast to domain library
