--- libs/co2/domain/src/lib/co2-domain.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  imports: [CommonModule],
-})
-export class Co2DomainModule {}
