+++ libs/co2/domain/src/lib/co2-forecast.ts
